Projeto limpo para reescrita das funções

+++ src/ProdutosDTO.java
@@ -10,7 +10,7 @@
 public class ProdutosDTO {
     private Integer id;
     private String nome;
-    private Double valor;
+    private int valor;
     private String status;
 
     public Integer getId() {
@@ -21,7 +21,7 @@ public class ProdutosDTO {
         return nome;
     }
 
-    public Double getValor() {
+    public int getValor() {
         return valor;
     }
 
@@ -37,7 +37,7 @@ public class ProdutosDTO {
         this.nome = nome;
     }
 
-    public void setValor(Double valor) {
+    public void setValor(int valor) {
         this.valor = valor;
     }
